refactor: NavHostViewModel stage 5 - extract drawer models and display mapping
